pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-ops-chat"
version = "0.1.0"
description = "Agentic RCA: logs → Chroma → Strands (Bedrock) → FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "chromadb>=0.5.0",
    "watchdog>=6.0.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "strands-agents>=1.30.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
